+++ src/api/DealsApi.js
@@ -29,6 +29,7 @@ import GetAddProductAttachementDetails from '../model/GetAddProductAttachementDe
 import GetAddedDeal from '../model/GetAddedDeal';
 import GetDeal from '../model/GetDeal';
 import GetDeals from '../model/GetDeals';
+import GetDealsCollection from '../model/GetDealsCollection';
 import GetDealsSummary from '../model/GetDealsSummary';
 import GetDealsTimeline from '../model/GetDealsTimeline';
 import GetDuplicatedDeal from '../model/GetDuplicatedDeal';
@@ -1610,6 +1611,94 @@ export default class DealsApi {
      */
     getDeals(opts) {
       return this.getDealsWithHttpInfo(opts)
+        .then(function(response_and_data) {
+          return response_and_data;
+        });
+    }
+
+
+    /**
+     * Get all deals (BETA)
+     * Returns all deals. This is a cursor-paginated endpoint that is currently in BETA. For more information, please refer to our documentation on <a href=\"https://pipedrive.readme.io/docs/core-api-concepts-pagination\" target=\"_blank\" rel=\"noopener noreferrer\">pagination</a>.
+     * @param {Object} opts Optional parameters
+     * @param {String} opts.cursor For pagination, the marker representing the first item on the next page. A Base64 encoded string containing `{\"deal\":id}`, e.g. `eyJhY3Rpdml0eSI6NDJ9`
+     * @param {Number} opts.limit For pagination, the limit of entries to be returned. If not provided, 100 items will be returned. Please note that a maximum value of 500 is allowed.
+     * @param {String} opts.since The time boundary that points to the start of the range of data. Datetime in ISO 8601 format. E.g. 2022-11-01 08:55:59. Operates on the `update_time` field.
+     * @param {String} opts.until The time boundary that points to the end of the range of data. Datetime in ISO 8601 format. E.g. 2022-11-01 08:55:59. Operates on the `update_time` field.
+     * @param {Number} opts.userId If supplied, only deals matching the given user will be returned
+     * @param {Number} opts.stageId If supplied, only deals within the given stage will be returned
+     * @param {module:model/String} opts.status Only fetch deals with a specific status. If omitted, all not deleted deals are returned. If set to deleted, deals that have been deleted up to 30 days ago will be included.
+     * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with an object containing data of type {@link module:model/GetDealsCollection} and HTTP response
+     */
+    getDealsCollectionWithHttpInfo(opts) {
+      opts = opts || {};
+      let postBody = null;
+
+
+
+
+
+
+
+
+      let pathParams = {
+      };
+      let queryParams = {
+        'cursor': opts['cursor'] === undefined ? opts['cursor'] : opts['cursor'],
+        'limit': opts['limit'] === undefined ? opts['limit'] : opts['limit'],
+        'since': opts['since'] === undefined ? opts['since'] : opts['since'],
+        'until': opts['until'] === undefined ? opts['until'] : opts['until'],
+        'user_id': opts['user_id'] === undefined ? opts['userId'] : opts['user_id'],
+        'stage_id': opts['stage_id'] === undefined ? opts['stageId'] : opts['stage_id'],
+        'status': opts['status'] === undefined ? opts['status'] : opts['status'],
+      };
+      let headerParams = {
+      };
+      let formParams = {
+      };
+
+      let formParamArray = [
+      ];
+
+      let contentTypes = [];
+      const isURLEncoded = contentTypes.includes('application/x-www-form-urlencoded');
+      const isJSON = contentTypes.includes('application/json');
+
+      if (isJSON) {
+        postBody = { ...postBody, ...opts };
+      } else if (isURLEncoded) {
+        for (let key in opts) {
+          if (opts.hasOwnProperty(key) && !formParamArray.includes(key)) {
+            formParams[key] = opts[key];
+          }
+        }
+      }
+
+      let authNames = ['api_key', 'oauth2', ];
+      let accepts = ['application/json', ];
+      let returnType = GetDealsCollection;
+      return this.apiClient.callApi(
+        '/deals/collection', 'GET',
+        pathParams, queryParams, headerParams, formParams, postBody,
+        authNames, contentTypes, accepts, returnType, null
+      );
+    }
+
+    /**
+     * Get all deals (BETA)
+     * Returns all deals. This is a cursor-paginated endpoint that is currently in BETA. For more information, please refer to our documentation on <a href=\"https://pipedrive.readme.io/docs/core-api-concepts-pagination\" target=\"_blank\" rel=\"noopener noreferrer\">pagination</a>.
+     * @param {Object} opts Optional parameters
+     * @param {String} opts.cursor For pagination, the marker representing the first item on the next page. A Base64 encoded string containing `{\"deal\":id}`, e.g. `eyJhY3Rpdml0eSI6NDJ9`
+     * @param {Number} opts.limit For pagination, the limit of entries to be returned. If not provided, 100 items will be returned. Please note that a maximum value of 500 is allowed.
+     * @param {String} opts.since The time boundary that points to the start of the range of data. Datetime in ISO 8601 format. E.g. 2022-11-01 08:55:59. Operates on the `update_time` field.
+     * @param {String} opts.until The time boundary that points to the end of the range of data. Datetime in ISO 8601 format. E.g. 2022-11-01 08:55:59. Operates on the `update_time` field.
+     * @param {Number} opts.userId If supplied, only deals matching the given user will be returned
+     * @param {Number} opts.stageId If supplied, only deals within the given stage will be returned
+     * @param {module:model/String} opts.status Only fetch deals with a specific status. If omitted, all not deleted deals are returned. If set to deleted, deals that have been deleted up to 30 days ago will be included.
+     * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with data of type {@link module:model/GetDealsCollection}
+     */
+    getDealsCollection(opts) {
+      return this.getDealsCollectionWithHttpInfo(opts)
         .then(function(response_and_data) {
           return response_and_data;
         });